Vendor SafeYAML as LessYAML
We vendor SafeYAML as LessYAML because we currently apply a monkey patch
to it that changes the behaviour of parsing certain types globally (e.g.
numbers are not turned into numbers, but kept as strings, strings like
"on" are not turned into booleans). This might affect YAML parsing in
apps loading this library (e.g. for ActiveRecord serialization or
whatever other YAML parsing that happens), and it is desirable to keep
this patch local to this library here.