Fix redundant job creation
When an expand key is included at the root, which it always is
thanks to the defaults being inserted at the parse stage (os),
the expansion code adds an extra job before checking the includes.

This removes that extra job and tries to do so as conservatively
as possible. There may be edge cases that this breaks, so
we should reconsider the defaults in the future.