[WIP] Web app (#9)
* Skeleton web app

Doesn't do much and isn't very robust, but here we go... :rocket:

* Add V1 and routing for future versions

* Extract auth to middleware

* Add messages to response

* Move route

* Simplify dirs and make Route available in higher namespace

* Home

* Handle parsing errors

* Enforce SSL in production

* Refactor route methods

* Add matrix to parse endpoint

* Dump JSON in strict mode (no symbols)

* Correct error class

* Split matrix expansion into separate endpoint

* Add both messages as attrs and full messages

Better for logging etc. to have the full messages, but some
clients would need to the separate message attrs.

Also adds .freeze in a bunch of places for perf.

* Symbolize keys before expansion

* Expand rows before decoration

* Add pry for debugging

* Add os to included rows if not present

* Normalize os array

* Add os to includes only is os is present

* Make matrix expansion more robust

Fixes:

1. nil values causing empty matrices (due to the way Array#product
   works)
2. Comparison of jobs in #uniq now compares all keys, even ones
   missing from some jobs

This also makes the matrix_spec.rb a proper unit test, which
does not rely on the default values and normalization that happens
when applying the spec.

* Fix include as hash regression

* Fix redundant job creation

When an expand key is included at the root, which it always is
thanks to the defaults being inserted at the parse stage (os),
the expansion code adds an extra job before checking the includes.

This removes that extra job and tries to do so as conservatively
as possible. There may be edge cases that this breaks, so
we should reconsider the defaults in the future.

* Make spec clearer

* Fix null include regression (517)

* add activesupport as prod dependency

* update gemfile.lock

* Unify alias reference

Because both keys and values can be aliased, we needed a name that
represents both in the alias info messsage. So it's now called
`actual`, and the message is:

"%{alias} is an alias for %{actual}, using %{actual}"

* Change default dist to trusty

* Allow unknown_key to be silenced with underscore

This allows users to define YAML aliases without getting an
unknown_key error. Example:

    _foo: &foo
      bar: 1
    <<: *foo

* Add CORS permissions for staging

This is to allow a tester app to make XHR requests.

There is no production use case for CORS at this time.

* Allow headers and post reqs

* add sentry

* remove ENV prod requirement