Pick the first var from env.jobs to jobs.include
Our system has always picked the first value from env.jobs to jobs
listed in jobs.include if they did not define a key env already.

This kind of makes sense for other matrix expansion keys, such as
rvm, which propagate to jobs.include in order to spare the user to
specify them over and over again.

For env vars it is pretty confusing behaviour, but it is what our
system has done for years, so we are porting it.

Might want to output a warning, and at some point deprecate this.