decouple defaults from required
so far defaults could only be defined on nodes (e.g. Language), not
mappings. therefore mappings would opt in to applying defaults by
marking a key required.

at the same time, in order to be able to make default values that depend
on their supporting context (such as dist: xenial, only valid on os:
linux) work, the semantic of "required" keys had been relaxed to mean
"required, but only if a value supported in the current context exists".

these semantic were slightly confusing, and have now changed, so that:

* a default is filled in on keys or nodes that define a default, and a
  supported value can be found, no matter if the key was marked required
* a required key will always produce an error message if no value was
  given

in return for these simplified, and more self-explaining semantics we
have to sacrifice the ability to correct an unknown value to a default
value because not all contexts now know these defaults. we could still
keep the behavior on nodes that have the defaults, but it makes more
sense to consistently just produce an error message saying that the
value is unknown.